refactor: organization of signup-related actions
